pyproject.toml
[project]
name = "decision-hub"
version = "0.1.0"
description = "The package manager & runtime for AI agent skills"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "typer[all]>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "boto3>=1.34.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "cryptography>=42.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "moto[s3]>=5.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
dhub = "decision_hub.cli.app:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/decision_hub"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
